MAINT: implement separate `backward_pass` method
